Redesigned and renamed the bot file to bot.js instead
