+++ package.json
@@ -2,7 +2,7 @@
   "name": "rinsedbot",
   "version": "1.0.0",
   "description": "A Discord bot to manage our server",
-  "main": "RinsedBot.js",
+  "main": "Bot.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
